+++ src/main/java/com/blackjack/util/NameParser.java
@@ -5,7 +5,7 @@ import static java.util.stream.Collectors.*;
 import java.util.List;
 import java.util.stream.Stream;
 
-import com.blackjack.domain.Name;
+import com.blackjack.domain.user.Name;
 
 public class NameParser {
 	private static final String DELIMITER = ",";
