+++ src/test/java/com/blackjack/domain/user/NameTest.java
@@ -1,4 +1,4 @@
-package com.blackjack.domain;
+package com.blackjack.domain.user;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+
+import com.blackjack.domain.user.Name;
 
 public class NameTest {
 	@DisplayName("올바른 값을 넣었을때 인스턴스 생성")
